requirements.txt
 
openai
langchain
PyPDF2
streamlit
python-dotenv
yt-dlp
assemblyai
langchain
chromadb
tiktoken